requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0
openpyxl>=3.1.0

# LLM Integration
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
langchain-core>=0.1.0
langgraph-prebuilt>=1.0.0

# UI (Optional - for Streamlit frontend)
streamlit>=1.28.0
plotly>=5.17.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Vector Store (Optional for RAG)
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Utilities
python-dateutil>=2.8.2
pydantic>=2.0.0

